requirements.txt
requests
pynacl
python-dotenv
cryptography